project cards added (may need some changes)
